Copy over revisions from the trunk:
  ------------------------------------------------------------------------
  r44790 | alagoutte | 2012-09-06 01:55:16 -0700 (Thu, 06 Sep 2012) | 16 lines
  Changed paths:
     M /trunk/epan/dissectors/packet-ieee80211.c

  From Masashi Honma via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7691
  TKIP dissection : wrong IS_TKIP macro

  In [1] "11.4.2.2 TKIP MPDU formats", we could see below sentence.
  "WEPSeed[1] is not used to construct the TSC, but is set to (TSC1 | 0x20) &
  0x7f."

  But the IS_TKIP macro only checks (WEPSeed[1] & 0x20).
  So sometimes IS_TKIP macro mis-dissects a CCMP frame as a TKIP frame.

  This patch changes IS_TKIP macro to do more better check.

  [1] IEEE Std 802.11.-2012

  #BACKPORT(1.8, 1.6)
  ------------------------------------------------------------------------
  r44844 | morriss | 2012-09-10 08:18:17 -0700 (Mon, 10 Sep 2012) | 6 lines
  Changed paths:
     M /trunk/epan/emem.c

  Attempt to fix https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5699
  and https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5979 :

  Don't cast the buffer address to int on Win32: a 64-bit address won't
  fit in an integer.
  ------------------------------------------------------------------------
  r44880 | morriss | 2012-09-11 06:11:33 -0700 (Tue, 11 Sep 2012) | 4 lines
  Changed paths:
     M /trunk/epan/emem.c

  As suggested by Anatoly in https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5699 :

  Followup to r44844: use a better cast (intptr_t instead of long long).
  ------------------------------------------------------------------------


Copy over with manual intervention:

  ------------------------------------------------------------------------
  r44881 | pascal | 2012-09-12 03:21:00 -0700 (Wed, 12 Sep 2012) | 3 lines
  Changed paths:
     M /trunk/epan/dissectors/packet-gtpv2.c

  From Esa Haapamäki via https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=7718 :
  Fix dissection of NAS Uplink and Downlink Count in MM Context IE
  ------------------------------------------------------------------------
  r44940 | etxrab | 2012-09-17 07:51:08 -0700 (Mon, 17 Sep 2012) | 1 line
  Changed paths:
     M /trunk/ui/gtk/voip_calls.c

  callsinfo->stop_fd not set in all branches for UNISTIM.
  ------------------------------------------------------------------------
  r45007 | cmaynard | 2012-09-19 10:57:54 -0700 (Wed, 19 Sep 2012) | 2 lines
  Changed paths:
     M /trunk/epan/dissectors/packet-icmp.c
     M /trunk/epan/dissectors/packet-icmpv6.c

  Fix request/response tracking with checksum corner case.
  ------------------------------------------------------------------------


Update the release notes.

svn path=/trunk-1.6/; revision=45235

+++ epan/emem.c
@@ -701,8 +701,8 @@ emem_create_chunk_gp(size_t size)
 	buf_end = npc->buf + size;
 
 	/* Align our guard pages on page-sized boundaries */
-	prot1 = (char *) ((((int) npc->buf + pagesize - 1) / pagesize) * pagesize);
-	prot2 = (char *) ((((int) buf_end - (1 * pagesize)) / pagesize) * pagesize);
+	prot1 = (char *) ((((intptr_t) npc->buf + pagesize - 1) / pagesize) * pagesize);
+	prot2 = (char *) ((((intptr_t) buf_end - (1 * pagesize)) / pagesize) * pagesize);
 
 	ret = VirtualProtect(prot1, pagesize, PAGE_NOACCESS, &oldprot);
 	g_assert(ret != 0 || versinfo.dwPlatformId == VER_PLATFORM_WIN32_WINDOWS);
